Add extract metadata to document
