+++ nesis/api/core/services/task_service.py
@@ -23,6 +23,8 @@ from nesis.api.core.models.entities import (
     Action,
     Datasource,
     Task,
+)
+from nesis.api.core.models.objects import (
     DatasourceStatus,
 )
 from nesis.api.core.models.objects import TaskType, TaskStatus
