+++ nesis/api/tests/core/document_loaders/test_s3.py
@@ -17,9 +17,12 @@ from nesis.api.core.models import DBSession
 from nesis.api.core.models import initialize_engine
 from nesis.api.core.models.entities import (
     Datasource,
+    Document,
+)
+
+from nesis.api.core.models.objects import (
     DatasourceType,
     DatasourceStatus,
-    Document,
 )
 from nesis.api.core.util import http
 from nesis.api.core.util.dateutil import strptime
