+++ nesis/api/tests/core/document_loaders/test_samba.py
@@ -14,6 +14,9 @@ from nesis.api.core.models import DBSession
 from nesis.api.core.models import initialize_engine
 from nesis.api.core.models.entities import (
     Datasource,
+)
+
+from nesis.api.core.models.objects import (
     DatasourceType,
     DatasourceStatus,
 )
